+++ src/hr/avrbanac/openglplayground/Globals.java
@@ -5,7 +5,7 @@ package hr.avrbanac.openglplayground;
  * Single place to hold all graphic globals.
  * 
  * @author avrbanac
- * @version 1.0.1
+ * @version 1.0.6
  */
 public class Globals {
     public static final int DISPLAY_WIDTH           = 1280;
@@ -18,9 +18,9 @@ public class Globals {
     public static final String TERRAIN_FRAGMENT_FILE    = "src/hr/avrbanac/openglplayground/shaders/terrainShader.frag";
     
     public static final String TEXTURE_TYPE         = "PNG";
-    public static final String TEXTURE_FOLDER       = "res/";
+    public static final String TEXTURE_FOLDER       = "res/texture/";
     public static final String TEXTURE_EXTENSION    = ".png";
-    public static final String MODEL_FOLDER         = "res/";
+    public static final String MODEL_FOLDER         = "res/model/";
     public static final String MODEL_EXTENSION      = ".obj";
     
     public static final float FOV                   = 70f;