+++ src/hr/avrbanac/openglplayground/textures/TerrainTexture.java
@@ -1,0 +1,21 @@
+
+package hr.avrbanac.openglplayground.textures;
+
+/**
+ * Texture terrain class.
+ * 
+ * @author avrbanac
+ * @version 1.0.6
+ */
+public class TerrainTexture {
+    private int textureID;
+
+    public TerrainTexture(int textureID) {
+        this.textureID = textureID;
+    }
+
+    public int getTextureID() {
+        return textureID;
+    }
+    
+}