+++ src/hr/avrbanac/openglplayground/textures/TerrainTexturePack.java
@@ -1,0 +1,39 @@
+
+package hr.avrbanac.openglplayground.textures;
+
+/**
+ * Terrain texture pack class.
+ * 
+ * @author avrbanac
+ * @version 1.0.6
+ */
+public class TerrainTexturePack {
+    private TerrainTexture backgroundTexture;
+    private TerrainTexture rTexture;
+    private TerrainTexture gTexture;
+    private TerrainTexture bTexture;
+
+    public TerrainTexturePack(TerrainTexture backgroundTexture, TerrainTexture rTexture, TerrainTexture gTexture, TerrainTexture bTexture) {
+        this.backgroundTexture = backgroundTexture;
+        this.rTexture = rTexture;
+        this.gTexture = gTexture;
+        this.bTexture = bTexture;
+    }
+
+    public TerrainTexture getBackgroundTexture() {
+        return backgroundTexture;
+    }
+
+    public TerrainTexture getrTexture() {
+        return rTexture;
+    }
+
+    public TerrainTexture getgTexture() {
+        return gTexture;
+    }
+
+    public TerrainTexture getbTexture() {
+        return bTexture;
+    }
+    
+}